block id in vote could be empty
