+++ tendermint/tests/support/rpc/block_empty_block_id.json
@@ -1,0 +1,120 @@
+{
+  "jsonrpc": "2.0",
+  "id": "",
+  "result": {
+    "block_meta": {
+      "block_id": {
+        "hash": "17E077714CE817C0DB3118D21EA060B2A16D9F07CB08AB603292F797FFD7F361",
+        "parts": {
+          "total": "1",
+          "hash": "1ED5B68A21D82D091C25D535F52C2D47419162618245C60D69B162A366F97A45"
+        }
+      },
+      "header": {
+        "version": {
+          "block": "10",
+          "app": "0"
+        },
+        "chain_id": "test-ab",
+        "height": "20",
+        "time": "2019-12-24T11:03:18.276852105Z",
+        "num_txs": "0",
+        "total_txs": "0",
+        "last_block_id": {
+          "hash": "08150B2F0F9D873C5446D77BF10EF1FD33A6C3FDE03BB76CC99F3A047A7A93CB",
+          "parts": {
+            "total": "1",
+            "hash": "EF93CDE980C98BD4720ADA7984121B4E0E153B004BB94E583809281C576FFAB1"
+          }
+        },
+        "last_commit_hash": "4D58DEA6B78D344255F6A7B0756F18D98DDFDF870D930337E13B6C7236BD07B0",
+        "data_hash": "",
+        "validators_hash": "A3E1386AC4A481A2A9DEB063B4D85973F1B789A2ECB15730B7E77FFFC698EC01",
+        "next_validators_hash": "A3E1386AC4A481A2A9DEB063B4D85973F1B789A2ECB15730B7E77FFFC698EC01",
+        "consensus_hash": "048091BC7DDC283F77BFBF91D73C44DA58C3DF8A9CBC867405D8B7F3DAADA22F",
+        "app_hash": "98F00A4EA270197259504954450D5ED5FA878284920FCC08070385C06F9B6515",
+        "last_results_hash": "",
+        "evidence_hash": "",
+        "proposer_address": "F18E6EE0D8AD707CED69D0606F30035B07E37DDF"
+      }
+    },
+    "block": {
+      "header": {
+        "version": {
+          "block": "10",
+          "app": "0"
+        },
+        "chain_id": "test-ab",
+        "height": "20",
+        "time": "2019-12-24T11:03:18.276852105Z",
+        "num_txs": "0",
+        "total_txs": "0",
+        "last_block_id": {
+          "hash": "08150B2F0F9D873C5446D77BF10EF1FD33A6C3FDE03BB76CC99F3A047A7A93CB",
+          "parts": {
+            "total": "1",
+            "hash": "EF93CDE980C98BD4720ADA7984121B4E0E153B004BB94E583809281C576FFAB1"
+          }
+        },
+        "last_commit_hash": "4D58DEA6B78D344255F6A7B0756F18D98DDFDF870D930337E13B6C7236BD07B0",
+        "data_hash": "",
+        "validators_hash": "A3E1386AC4A481A2A9DEB063B4D85973F1B789A2ECB15730B7E77FFFC698EC01",
+        "next_validators_hash": "A3E1386AC4A481A2A9DEB063B4D85973F1B789A2ECB15730B7E77FFFC698EC01",
+        "consensus_hash": "048091BC7DDC283F77BFBF91D73C44DA58C3DF8A9CBC867405D8B7F3DAADA22F",
+        "app_hash": "98F00A4EA270197259504954450D5ED5FA878284920FCC08070385C06F9B6515",
+        "last_results_hash": "",
+        "evidence_hash": "",
+        "proposer_address": "F18E6EE0D8AD707CED69D0606F30035B07E37DDF"
+      },
+      "data": {
+        "txs": null
+      },
+      "evidence": {
+        "evidence": null
+      },
+      "last_commit": {
+        "block_id": {
+          "hash": "08150B2F0F9D873C5446D77BF10EF1FD33A6C3FDE03BB76CC99F3A047A7A93CB",
+          "parts": {
+            "total": "1",
+            "hash": "EF93CDE980C98BD4720ADA7984121B4E0E153B004BB94E583809281C576FFAB1"
+          }
+        },
+        "precommits": [
+          {
+            "type": 2,
+            "height": "19",
+            "round": "0",
+            "block_id": {
+              "hash": "",
+              "parts": {
+                "total": "0",
+                "hash": ""
+              }
+            },
+            "timestamp": "2019-12-24T11:03:18.534146209Z",
+            "validator_address": "9ACE9BC3DC3455BC89427F53172077D84E54D5BD",
+            "validator_index": "0",
+            "signature": "1Go3/vHT/g9tplm3TtaiFupZG1tYYycLf4HXbVJvN3lWD91KwZA4SNKSfzRX5ccOQXy1QTqjLk1Zquc/4w00DQ=="
+          },
+          {
+            "type": 2,
+            "height": "19",
+            "round": "0",
+            "block_id": {
+              "hash": "08150B2F0F9D873C5446D77BF10EF1FD33A6C3FDE03BB76CC99F3A047A7A93CB",
+              "parts": {
+                "total": "1",
+                "hash": "EF93CDE980C98BD4720ADA7984121B4E0E153B004BB94E583809281C576FFAB1"
+              }
+            },
+            "timestamp": "2019-12-24T11:03:18.276852105Z",
+            "validator_address": "F18E6EE0D8AD707CED69D0606F30035B07E37DDF",
+            "validator_index": "1",
+            "signature": "pE4CEh2HnXmZ8UfCcvuQ/3udpzoWTT9cxRzEv5XdTyV+WEQNna8oO2uflFc/5wORB/DoGmUjOlfS512HcfFuBQ=="
+          }
+        ]
+      }
+    }
+  }
+}
